Don't crash when a controller sends invalid log entries.
A log entry with an unknown tag ID is disallowed by SQL constraints, but can occur when the controller has an outdated local database or when tags were deleted.

Invalid input should be handled by a corresponding HTTP response, not by crashing with an uncaught exception. (There are a lot more such suspicious throws around, I'm not fixing all of them here!) However, in this case that wouldn't help either, because the controller would just try to send the invalid entry again later and get stuck on it. Instead, accept it and log all the relevant information to troubleshoot the problem.